pyproject.toml
[tool.poetry]
name = "langchain-lab"
version = "0.1.0"
description = ""
authors = ["Lei Zhang <zhanglei@apache.org>"]
license = "Apache License 2.0"
readme = "README.md"
packages = [{ include = "langchain_lab", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.27.0"
langchain = "0.1.4"
langchain-openai = "0.0.5"
langchain-experimental = "0.0.49"
faiss-cpu = "^1.7.4"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"
docx2txt = "^0.8"
pymupdf = "^1.23.3"
sentence-transformers = "^2.2.2"
langdetect = "^1.0.9"
chardet = "^5.2.0"
beautifulsoup4 = "^4.12.2"
SQLAlchemy = "2.0.22"

[tool.pytest.ini_options]
pythonpath = ["src"]

[tool.poetry.group.lint.dependencies]
isort = "^5.12.0"
black = { version = "^23.1a1", allow-prereleases = true }
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 180

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
priority = "supplemental"

[[tool.poetry.source]]
name = "aliyun"
url = "https://mirrors.aliyun.com/pypi/simple"
priority = "primary"
